Max#9
--Issues--
-Fixed a bug where a human could trigger monster commands.
-Cursor enabling and disabling now work more accurately.

--Monster--
-Added a menu for the monster.
-Added a HUD for the monster.
-Added Idle and Move animations for the monster.

--Other--
Most bugs in the lobby have been fixed, including the ability to return from the game.
A custom lobby has been added, which can be modified as desired.
Added the ability to see players and the monster during selection.
The monster selection button has been fixed: reconnecting to the lobby now correctly displays the button as disabled if the monster is already selected by someone. The only remaining task is to handle this during the initial connection to the server.